Moved gamegrid call to prevent walking through walls
